Fix: separate proxy dependencies to modules
